bsp: clean up
